move interfaces to one file
