+++ client/src/components/contexts/filteredJobs.tsx
@@ -1,8 +1,6 @@
 // Package imports
 import React, { createContext,Dispatch, SetStateAction, useContext, useState } from 'react';
-import { job } from '../../interfaces'
-import { Children } from './interfaces';
-
+import { job, Children } from '../../interfaces'
 
 const Context = createContext<[job[], Dispatch<SetStateAction<job[]>>] | null>(null);
