--- client/src/components/contexts/interfaces.ts
@@ -1,1 +1,0 @@
-export interface Children { children: React.ReactNode }
