+++ client/src/components/contexts/jobs.tsx
@@ -1,7 +1,5 @@
 import React, { createContext,Dispatch, SetStateAction, useContext, useState } from 'react';
-import { job } from '../../interfaces'
-import { Children } from './interfaces';
-
+import { job, Children } from '../../interfaces'
 
 const Context = createContext<[job[], Dispatch<SetStateAction<job[]>>] | null>(null);
